Only show Delete on New Responses screen (#2871)
* only show delete on New

* lift delete button up

* Extract Delete to component and only show on New
